적 AI 업그레이드 / 적 공격 로직 추가 / 플레이어 피격 기능 제작 중

+++ Assets/FPSGame/Scripts/Enemy/EnemyData.cs
@@ -26,7 +26,7 @@ namespace FPSGame
         public float FireRate { get { return fireRate; } }
 
         [SerializeField] private int maxBullet = 3;                   // 최대 탄약 수.
-        public float MaxBullet { get { return maxBullet; } }
+        public int MaxBullet { get { return maxBullet; } }
 
         [SerializeField] private float maxHP = 100f;                  // 체력
         public float MaxHP { get { return maxHP; } }
